pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "humanatlas-dashboard"
version = "0.1.0"
description = "Dashboard and analysis pipeline for Human Atlas interaction telemetry."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "duckdb>=1.1.0",
  "pandas>=2.2.0",
  "plotly>=5.24.0",
  "pyarrow>=17.0.0",
  "pyyaml>=6.0.2",
  "streamlit>=1.39.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]
deliverables = [
  "python-pptx>=1.0.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
